requirements.txt
streamlit
pandas
mlxtend
mysql-connector-python
openpyxl